chore: Update wait_random function to use int type for max_delay parameter

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,7 +4,7 @@ import random
 import asyncio
 
 
-async def wait_random(max_delay: float = 10) -> float:
-    wait = random.uniform(0.0, max_delay)
+async def wait_random(max_delay: int = 10) -> float:
+    wait = random.uniform(0, max_delay)
     await asyncio.sleep(wait)
     return wait
